Switched from TCP to UDP, can now continue flights
Switched to TCP and UDP as there were flushing issues
This applies to the sending of data from Python to Node.js.
Node.js -> Python is still TCP, no observable issues yet

Can now press a button to continue flying. This allows for landing
and not changing the battery but still keeping track of how long
one was flying for

Fixed a bug where the timer was displaying the last two digits of
a number like '08.183848' instead of '08'. As everything was in
seconds except this, it stayed a static '48' for the entire flight